+++ storybook/src/components/legend/legend-item/legend-item.decorator.ts
@@ -1,5 +1,8 @@
+import { TemplateResult, html } from "lit-html";
+import { PartialStoryFn } from "storybook/internal/types";
+
 // https://github.com/dso-toolkit/dso-toolkit/issues/1313#issue-1041224938
-export const legendItemDemoCss = `
+const legendItemDemoCss = `
   .symboolcode[data-symboolcode=regelingsgebied] {
     display: block;
     position: absolute;
@@ -7,4 +10,12 @@ export const legendItemDemoCss = `
     background-image: url(images/regelingsgebied.png);
     background-size: contain;
   }
+`;
+
+export const decorator = (story: PartialStoryFn): TemplateResult => html`
+  ${story()}
+
+  <style>
+    ${legendItemDemoCss}
+  </style>
 `;